Make an implicit assumption about data size explicit in some tests
Immutable shares are limited to what you can pay for with 10 passes right
now.  This certainly needs to be lifted but until it is we don't want spurious
test failures telling us about the limit.

Mutable shares are limited to what you can pay for with 1 pass right now
because the client doesn't know how to compute the right number of passes and
always spends 1 pass.