A lot of fixes
The native st_size-based approach to size estimation is broken because mutable
and immutable layouts have a different amount of overhead.  So start peeking
inside the share files.

Also, new_length must be respected now because when it is used the part of the
share file we peek at changes.  Also, new_length can *only* truncate so stop
treating it as if it can extend.