Get rid of this do-nothing function
It was doing to do some complex time math and decide if the voucher redemption
looked legit but the whole premise is broken.

The only thing you might be able to decide about a key change on a new voucher
is that some *older* ZKAPs might be suspicious.  And, really, what you should
probably decide instead is that whole issuer is suspicious and you might need
to stop using it.  On top of that, you need more information than what you get
from voucher redemptions.  You need more frequent monitoring of the issuer's
key in case you don't redeem vouchers except once in a blue moon.

So hanging a suspiciousness on a voucher isn't coherent.