Go back to regular resource class
We collect coverage info so we can't very easily run on multiple cores.  So we
mostly can't take advantage of the bigger machine.

With Cachix we can now cache even partial build successes so even if we lose
the whole cache and have to start from scratch, the build should now converge
on success even if it takes a few tries.