Make sure that test_renewed is exercising the interesting code path
Previously Hypothesis spent a long while putting storage indexes on the
storage servers and creating a list of nodes with non-overlapping storage
indexes and then asserting that, out of no relevant shares, none had expired
leases.

Now all storage servers in the test will have all of the relevant shares and
so all should be visited and updated.