Change implementer_only to implementer.
This removes the need to call `registerAdapter` on
`ZKAPAuthorizerStorageServer`. It looks like that was needed because the
interfaces for it were defined by `implementer_only` which explicitly ignored
interfaces defined by parent classes. Looking at the history, I assume the use
of `implementer_only` was to avoid implementing `RIStorageServer` via
`proxyForInterface`. Given that the later is no longer used, `implementer_only`
is no longer necessary.