Replace versioneer with hand-rolled version generation code.
The version of versioneer has an issue that the tarball that is generated on
github changes, when the branches and/or tags refering to a commit change. This
causes issues for gridsync, which pins the hashes of those tarballs.

In addition to that, versioneer doesn't provide a meaningful revision for
untagged commits, when run from a github tarball.

Given that we only ever consume a github tarball as a github archive (except in
local development, where the version number is less important), this replaces
versioneer with some hand-rolled code specialized to that use case.