use the ULP-distance-based matcher to match recovered floats
In this commit the distance is constrained to 0 which is the same as exact
value comparison.  I want to see what happens on other platforms with this
configuration.  In a subsequent commit I will probably raise it to 4 which
is the largest distance found by Rick Regan and written up on
<https://www.exploringbinary.com/incorrect-decimal-to-floating-point-conversion-in-sqlite/>.
He makes no guarantee that this is the greatest possible error but it seems
like a reasonable next test.