Improve typing around MemoryGrid
This gets rid of the ability to download a bespoke JSON encoding of a
directory from MemoryGrid, which the real Tahoe-LAFS client does not offer.

It also gets rid of a test that relied on this.  However, there is another
test that sufficiently exercises `make_directory` by using `list_directory`
instead.