fix failing test_replicating
First drop the test helper and just define two tests.  Should have commit this
separately, oops.

Second, re-add the persistent strong reference to `store` in
test_replicating.  If we don't keep the store alive for the whole test then it
is garbage collected and `_get_store` builds us a new one - and its connection
does not match the connection held by the replication service.

Another nail in the coffin of the weakref caching scheme...