expose recovery capability from setup endpoint even for CONFLICT case
If a replica is already set up then be sure to return it in the CONFLICT error
response if a client tries to set up replication again.

Also improve the OpenAPI spec by deleting the stuff that no longer exists,
update the design doc to reflect the websocket-based API, and make a few mypy
improvements in the area of the changing code.