Bump mach-nix for some bug fixes and pypi-deps-db for new packages
mach-nix 3.5.0 has problems resolving some Python 3.10 wheels but this new
master revision solves them.  This update also fixes the problem resolving the
`flit-core` dependency but I don't know why.

Also bump pypi-deps-db so we have a version that includes the most recent
python-challenge-bypass-ristretto release.  This is probably not strictly
necessary but it's not a bad thing to let us update that package.